requirements.txt
opencv-python
compreface-sdk
grpcio-tools
yolov5

# IMPORTANT: Uninstall the following packages before installing the new ones
# pip uninstall requests_toolbelt
# pip uninstall urllib3
requests_toolbelt
urllib3